+++ src/multiwinner/randballots.h
@@ -20,8 +20,8 @@ class random_ballots : public multiwinner_method {
 		mutable rng random_gen;
 
 	public:
-		std::list<int> get_council(int council_size,
-			int num_candidates, const election_t & ballots) const;
+		std::list<size_t> get_council(size_t council_size,
+			size_t num_candidates, const election_t & ballots) const;
 
 		std::string name() const {
 			return "Random ballots";
